Refactoring to have test helper in .spec.ts
